// shared/enums.ts
export enum Role {
  ADMIN = 'admin',
  GUEST = 'guest',
  EMPLOYER = 'employer',
  EMPLOYEE = 'employee'
}

export enum LogReason {
  MONGODB = 'mongodb',
  VALIDATION = 'validation'
}
export enum Collections {
  MONGO_DB_USER_COLLECTION_NAME = 'users',
  MONGO_DB_JOBS_COLLECTION_NAME = 'jobs',
  MONGO_DB_NEGOTIATION_COLLECTION_NAME = 'negotiation'
}